pyproject.toml
[tool.poetry]
name = "cyber-guardians-api"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = { extras = ["standard"], version = "^0.115.4" }
motor = "^3.6.0"
pydantic-settings = "^2.6.1"
pytest-asyncio = "^0.24.0"
pyjwt = "^2.9.0"
bcrypt = "^4.2.0"
cryptography = "^43.0.3"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-cov = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
